Move Notif handling to MainActivity
So it doesnt explode when you swap fragments...
